Move name-dependent methods to ShowNameHelper
Move from ShowObservationHelper.
They depend on, and show, info about the name, not the obs.
