+++ app/controllers/application_controller.rb
@@ -122,8 +122,8 @@ class ApplicationController < ActionController::Base
   # after_action  :extra_gc
   # after_action  :log_memory_usage
 
-  # Make show_observation_helper available to nested partials
-  helper :show_observation
+  # Make show_name_helper available to nested partials
+  helper :show_name
 
   # Disable all filters except set_locale.
   # (Used to streamline API and Ajax controllers.)
